# app/src/lib/api/generated/openAPI.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * OpenAPI
 * 
This is the autogenerated OpenAPI documentation for your [Fuego](https://github.com/go-fuego/fuego) API.

Below is a Fuego Cheatsheet to help you get started. Don't hesitate to check the [Fuego documentation](https://go-fuego.dev) for more details.

Happy coding! 🔥

## Usage

### Route registration

```go
func main() {
	// Create a new server
	s := fuego.NewServer()

	// Register some routes
	fuego.Post(s, "/hello", myController)
	fuego.Get(s, "/myPath", otherController)
	fuego.Put(s, "/hello", thirdController)

	adminRoutes := fuego.Group(s, "/admin")
	fuego.Use(adminRoutes, myMiddleware) // This middleware (for authentication, etc...) will be available for routes starting by /admin/*, 
	fuego.Get(adminRoutes, "/hello", groupController) // This route will be available at /admin/hello

	// Start the server
	s.Start()
}
```

### Basic controller

```go
type MyBody struct {
	Name string `json:"name" validate:"required,max=30"`
}

type MyResponse struct {
	Answer string `json:"answer"`
}

func hello(ctx fuego.ContextWithBody[MyBody]) (*MyResponse, error) {
	body, err := ctx.Body()
	if err != nil {
		return nil, err
	}

	return &MyResponse{Answer: "Hello " + body.Name}, nil
}
```

### Add openAPI information to the route

```go
import (
	"github.com/go-fuego/fuego"
	"github.com/go-fuego/fuego/option"
	"github.com/go-fuego/fuego/param"
)

func main() {
	s := fuego.NewServer()

	// Custom OpenAPI options
	fuego.Post(s, "/", myController
		option.Description("This route does something..."),
		option.Summary("This is my summary"),
		option.Tags("MyTag"), // A tag is set by default according to the return type (can be deactivated)
		option.Deprecated(), // Marks the route as deprecated in the OpenAPI spec

		option.Query("name", "Declares a query parameter with default value", param.Default("Carmack")),
		option.Header("Authorization", "Bearer token", param.Required()),
		optionPagination,
		optionCustomBehavior,
	)

	s.Run()
}

var optionPagination = option.Group(
	option.QueryInt("page", "Page number", param.Default(1), param.Example("1st page", 1), param.Example("42nd page", 42)),
	option.QueryInt("perPage", "Number of items per page"),
)

var optionCustomBehavior = func(r *fuego.BaseRoute) {
	r.XXX = "YYY"
}
```

Then, in the controller

```go
type MyResponse struct {
	Answer string `json:"answer"`
}

func getAllPets(ctx fuego.ContextNoBody) (*MyResponse, error) {
	name := ctx.QueryParam("name")
	perPage, _ := ctx.QueryParamIntErr("per_page")

	return &MyResponse{Answer: "Hello " + name}, nil
}
```

 * OpenAPI spec version: 0.0.1
 */
/**
 * CreateRoomRequest schema
 */
export interface CreateRoomRequest {
	/**
	 * @minLength 1
	 * @maxLength 50
	 */
	name: string;
	/**
	 * @minLength 4
	 * @maxLength 72
	 */
	password?: string | null;
}

/**
 * Additional information about the error
 */
export type HTTPErrorErrorsItemMore = { [key: string]: unknown | null } | null;

export type HTTPErrorErrorsItem = {
	/** Additional information about the error */
	more?: HTTPErrorErrorsItemMore;
	/** For example, name of the parameter that caused the error */
	name?: string;
	/** Human readable error message */
	reason?: string;
} | null;

/**
 * HTTPError schema
 */
export interface HTTPError {
	/** Human readable error message */
	detail?: string | null;
	errors?: HTTPErrorErrorsItem[] | null;
	instance?: string | null;
	/** HTTP status code */
	status?: number | null;
	/** Short title of the error */
	title?: string | null;
	/** URL of the error type. Can be used to lookup the error in a documentation */
	type?: string | null;
}

/**
 * Message schema
 */
export interface Message {
	content?: string;
	id?: string;
	pubkey?: string | null;
	room?: string;
	signature?: string | null;
	signed_timestamp?: number | null;
	timestamp?: string;
	user?: string;
}

/**
 * Room schema
 */
export interface Room {
	has_password?: boolean;
	last_message_content?: string | null;
	last_message_timestamp?: string | null;
	last_message_user?: string | null;
	name?: string;
}

/**
 * SendMessageRequest schema
 */
export interface SendMessageRequest {
	content: string;
	pubkey?: string | null;
	room_password?: string | null;
	signature?: string | null;
	timestamp?: number | null;
	user: string;
}

/**
 * ServerInfoResponse schema
 */
export interface ServerInfoResponse {
	description?: string;
	suggested_quickname?: string;
}

/**
 * User schema
 */
export interface User {
	created_at?: string;
	public_key?: string;
	updated_at?: string;
	verified?: boolean;
}

/**
 * unknown-interface schema
 */
export interface UnknownInterface {}

export type GETApiRoomsParams = {
	visited?: string;
};

export type GETApiRoomsSearchParams = {
	visited?: string;
	q?: string;
};

export type GETApiRoomsRoomMessagesParams = {
	password?: string;
};